Split recursive key delta execution into its own file
