+++ src/execution/operator/set/physical_recursive_cte.cpp
@@ -1,8 +1,8 @@
 #include "duckdb/execution/operator/set/physical_recursive_cte_state.hpp"
+#include "duckdb/execution/operator/set/physical_recursive_cte_delta.hpp"
 
 #include "duckdb/execution/operator/scan/physical_column_data_scan.hpp"
 #include "duckdb/common/vector/flat_vector.hpp"
-#include "duckdb/common/vector_operations/vector_operations.hpp"
 #include "duckdb/main/client_context.hpp"
 #include "duckdb/parallel/pipeline_executor.hpp"
 #include "duckdb/storage/buffer_manager.hpp"
@@ -26,64 +26,6 @@ struct RecursiveCTEDistinctPartition {
 
 	mutex lock;
 	GroupedAggregateHashTable ht;
-};
-
-struct RecursiveCTEKeyDeltaState {
-	RecursiveCTEKeyDeltaState(ClientContext &context, const PhysicalRecursiveCTE &op)
-	    : touched_keys(context, BufferAllocator::Get(context), op.distinct_types),
-	      previous_rows(context, op.GetTypes()), new_keys(context, op.distinct_types),
-	      touched_addresses(LogicalType::POINTER), matched_addresses(LogicalType::POINTER),
-	      first_touches(STANDARD_VECTOR_SIZE), found_groups(STANDARD_VECTOR_SIZE), missing_groups(STANDARD_VECTOR_SIZE),
-	      changed_groups(STANDARD_VECTOR_SIZE), changed_column_groups(STANDARD_VECTOR_SIZE),
-	      equal_groups_a(STANDARD_VECTOR_SIZE), equal_groups_b(STANDARD_VECTOR_SIZE), arena(Allocator::Get(context)),
-	      row_state(arena) {
-		first_touch_keys.Initialize(Allocator::Get(context), op.distinct_types);
-		selected_keys.Initialize(Allocator::Get(context), op.distinct_types);
-		payload_rows.Initialize(Allocator::Get(context), op.payload_types);
-		result_rows.Initialize(Allocator::Get(context), op.GetTypes());
-		changed_rows.Initialize(Allocator::Get(context), op.GetTypes());
-		previous_scan_rows.Initialize(Allocator::Get(context), op.GetTypes());
-		key_scan_rows.Initialize(Allocator::Get(context), op.distinct_types);
-	}
-
-	void Reset() {
-		touched_keys.ResetForNewIteration(0);
-		previous_rows.ResetForReuse();
-		previous_rows.InitializeAppend(previous_append_state);
-		new_keys.ResetForReuse();
-		new_keys.InitializeAppend(new_key_append_state);
-		touched_count = 0;
-		new_count = 0;
-		changed_count = 0;
-	}
-
-	GroupedAggregateHashTable touched_keys;
-	ColumnDataCollection previous_rows;
-	ColumnDataCollection new_keys;
-	ColumnDataAppendState previous_append_state;
-	ColumnDataAppendState new_key_append_state;
-	DataChunk first_touch_keys;
-	DataChunk selected_keys;
-	DataChunk payload_rows;
-	DataChunk result_rows;
-	DataChunk changed_rows;
-	DataChunk previous_scan_rows;
-	DataChunk key_scan_rows;
-	Vector touched_addresses;
-	Vector matched_addresses;
-	SelectionVector first_touches;
-	SelectionVector found_groups;
-	SelectionVector missing_groups;
-	SelectionVector changed_groups;
-	SelectionVector changed_column_groups;
-	SelectionVector equal_groups_a;
-	SelectionVector equal_groups_b;
-	AggregateHTLookupState lookup_state;
-	ArenaAllocator arena;
-	RowOperationsState row_state;
-	idx_t touched_count = 0;
-	idx_t new_count = 0;
-	idx_t changed_count = 0;
 };
 
 PhysicalRecursiveCTE::PhysicalRecursiveCTE(PhysicalPlan &physical_plan, Identifier ctename, TableIndex table_index,
@@ -494,134 +436,6 @@ static void ScatterChunk(DataChunk &output_chunk, DataChunk &input_chunk, const 
 	idx_t chunk_index = 0;
 	for (auto &group_idx : idx_set) {
 		output_chunk.data[group_idx].Reference(input_chunk.data[chunk_index++]);
-	}
-}
-
-void RecursiveCTEState::SnapshotUsingKeyDelta(DataChunk &keys) {
-	D_ASSERT(key_delta);
-	auto &delta = *key_delta;
-	const auto first_touch_count =
-	    delta.touched_keys.FindOrCreateGroups(keys, delta.touched_addresses, delta.first_touches);
-	delta.touched_count += first_touch_count;
-	if (first_touch_count == 0) {
-		return;
-	}
-
-	delta.first_touch_keys.Reset();
-	delta.first_touch_keys.Slice(keys, delta.first_touches, first_touch_count);
-	const auto found_count = ht->LookupGroups(delta.first_touch_keys, delta.lookup_state, delta.found_groups);
-
-	idx_t found_idx = 0;
-	idx_t missing_count = 0;
-	for (idx_t key_idx = 0; key_idx < first_touch_count; key_idx++) {
-		if (found_idx < found_count && delta.found_groups.get_index_unsafe(found_idx) == key_idx) {
-			found_idx++;
-			continue;
-		}
-		delta.missing_groups.set_index(missing_count++, key_idx);
-	}
-	D_ASSERT(found_idx == found_count);
-	delta.new_count += missing_count;
-
-	if (found_count > 0) {
-		delta.selected_keys.Reset();
-		delta.selected_keys.Slice(delta.first_touch_keys, delta.found_groups, found_count);
-
-		delta.matched_addresses.SetVectorType(VectorType::FLAT_VECTOR);
-		auto source_addresses = FlatVector::GetData<data_ptr_t>(delta.lookup_state.addresses);
-		auto target_addresses = FlatVector::GetDataMutable<data_ptr_t>(delta.matched_addresses);
-		for (idx_t match_idx = 0; match_idx < found_count; match_idx++) {
-			target_addresses[match_idx] = source_addresses[delta.found_groups.get_index_unsafe(match_idx)];
-		}
-		FlatVector::SetSize(delta.matched_addresses, found_count);
-
-		delta.payload_rows.Reset();
-		delta.payload_rows.SetChildCardinality(found_count);
-		if (delta.payload_rows.ColumnCount() > 0) {
-			FinalizePayload(delta.row_state, delta.matched_addresses, delta.payload_rows, 0);
-		}
-		delta.result_rows.Reset();
-		ScatterChunk(delta.result_rows, delta.selected_keys, op.distinct_idx);
-		ScatterChunk(delta.result_rows, delta.payload_rows, op.payload_idx);
-		delta.result_rows.CheckCardinality(found_count);
-		delta.previous_rows.Append(delta.previous_append_state, delta.result_rows);
-	}
-
-	if (missing_count > 0) {
-		delta.selected_keys.Reset();
-		delta.selected_keys.Slice(delta.first_touch_keys, delta.missing_groups, missing_count);
-		delta.new_keys.Append(delta.new_key_append_state, delta.selected_keys);
-	}
-}
-
-void RecursiveCTEState::FinalizeUsingKeyDelta() {
-	D_ASSERT(key_delta);
-	auto &delta = *key_delta;
-	ColumnDataScanState previous_scan_state;
-	delta.previous_rows.InitializeScan(previous_scan_state);
-	while (delta.previous_rows.Scan(previous_scan_state, delta.previous_scan_rows)) {
-		const auto row_count = delta.previous_scan_rows.size();
-		delta.key_scan_rows.Reset();
-		GatherChunk(delta.key_scan_rows, delta.previous_scan_rows, op.distinct_idx);
-		const auto found_count = ht->LookupGroups(delta.key_scan_rows, delta.lookup_state, delta.found_groups);
-		if (found_count != row_count) {
-			throw InternalException("USING KEY delta finalization could not find %d of %d touched groups",
-			                        row_count - found_count, row_count);
-		}
-
-		delta.payload_rows.Reset();
-		delta.payload_rows.SetChildCardinality(row_count);
-		if (delta.payload_rows.ColumnCount() > 0) {
-			FinalizePayload(delta.row_state, delta.lookup_state.addresses, delta.payload_rows, 0);
-		}
-		delta.result_rows.Reset();
-		ScatterChunk(delta.result_rows, delta.key_scan_rows, op.distinct_idx);
-		ScatterChunk(delta.result_rows, delta.payload_rows, op.payload_idx);
-		delta.result_rows.CheckCardinality(row_count);
-
-		idx_t changed_count = 0;
-		idx_t equal_count = row_count;
-		optional_ptr<const SelectionVector> equal_groups;
-		for (idx_t payload_idx = 0; payload_idx < op.payload_idx.size() && equal_count > 0; payload_idx++) {
-			auto &next_equal_groups = payload_idx % 2 == 0 ? delta.equal_groups_a : delta.equal_groups_b;
-			const auto changed_column_count = VectorOperations::DistinctFrom(
-			    delta.previous_scan_rows.data[op.payload_idx[payload_idx]], delta.payload_rows.data[payload_idx],
-			    equal_groups, equal_count, &delta.changed_column_groups, &next_equal_groups);
-			for (idx_t changed_idx = 0; changed_idx < changed_column_count; changed_idx++) {
-				delta.changed_groups.set_index(changed_count++,
-				                               delta.changed_column_groups.get_index_unsafe(changed_idx));
-			}
-			equal_count -= changed_column_count;
-			equal_groups = &next_equal_groups;
-		}
-		if (changed_count > 0) {
-			delta.changed_count += changed_count;
-			delta.changed_rows.Reset();
-			delta.changed_rows.Slice(delta.result_rows, delta.changed_groups, changed_count);
-			op.working_table->Append(working_append_state, delta.changed_rows);
-		}
-	}
-
-	ColumnDataScanState new_key_scan_state;
-	delta.new_keys.InitializeScan(new_key_scan_state);
-	while (delta.new_keys.Scan(new_key_scan_state, delta.key_scan_rows)) {
-		const auto row_count = delta.key_scan_rows.size();
-		const auto found_count = ht->LookupGroups(delta.key_scan_rows, delta.lookup_state, delta.found_groups);
-		if (found_count != row_count) {
-			throw InternalException("USING KEY delta finalization could not find %d of %d new groups",
-			                        row_count - found_count, row_count);
-		}
-
-		delta.payload_rows.Reset();
-		delta.payload_rows.SetChildCardinality(row_count);
-		if (delta.payload_rows.ColumnCount() > 0) {
-			FinalizePayload(delta.row_state, delta.lookup_state.addresses, delta.payload_rows, 0);
-		}
-		delta.result_rows.Reset();
-		ScatterChunk(delta.result_rows, delta.key_scan_rows, op.distinct_idx);
-		ScatterChunk(delta.result_rows, delta.payload_rows, op.payload_idx);
-		delta.result_rows.CheckCardinality(row_count);
-		op.working_table->Append(working_append_state, delta.result_rows);
 	}
 }
